[REF] restructure project to new states
