+++ fastapi_babel/__init__.py
@@ -1,13 +1,18 @@
-from .core import Babel, BabelCli, _
+from .core import Babel
+from .helpers import _, LazyText as lazy_gettext, use_babel
+from .cli import BabelCli
+from .local_context import BabelContext
 from .middleware import BabelMiddleware
 from .properties import RootConfigs as BabelConfigs
 
-__version__ = "0.0.9"
+__version__ = "0.1.0"
 __author__ = "papuridalego@gmail.com"
 __all__ = [
     "Babel",
     "BabelCli",
     "BabelConfigs",
     "_",
+    "lazy_gettext",
+    "BabelContext",
     "BabelMiddleware",
 ]
