+++ fastapi_babel/properties.py
@@ -15,9 +15,23 @@ class RootConfigs:
     BABEL_MESSAGE_POT_FILE: str = field(init=False)
 
     def __post_init__(self):
-        self.ROOT_DIR = pathlib.Path(self.ROOT_DIR).parent
-        self.BABEL_TRANSLATION_DIRECTORY = os.path.join(
-            self.ROOT_DIR, self.BABEL_TRANSLATION_DIRECTORY
+        setattr(
+            self,
+            "ROOT_DIR",
+            pathlib.Path(self.ROOT_DIR).parent,
         )
-        self.BABEL_CONFIG_FILE = os.path.join(self.ROOT_DIR, self.BABEL_CONFIG_FILE)
-        self.BABEL_MESSAGE_POT_FILE = os.path.join(self.ROOT_DIR, self.BABEL_DOMAIN)
+        setattr(
+            self,
+            "BABEL_TRANSLATION_DIRECTORY",
+            os.path.join(self.ROOT_DIR, self.BABEL_TRANSLATION_DIRECTORY),
+        )
+        setattr(
+            self,
+            "BABEL_CONFIG_FILE",
+            os.path.join(self.ROOT_DIR, self.BABEL_CONFIG_FILE),
+        )
+        setattr(
+            self,
+            "BABEL_MESSAGE_POT_FILE",
+            os.path.join(self.ROOT_DIR, self.BABEL_DOMAIN),
+        )
